Design: 마이 페이지 디자인 수정
